webapp/requirements.txt
streamlit
python-dotenv
azure-iot-hub
psycopg2-binary